Implemented PopOverHelper changingView.
