+++ src/ch/imedias/rsccfx/view/RsccRequestPresenter.java
@@ -18,6 +18,7 @@ public class RsccRequestPresenter implements ControlledPresenter {
   private final RsccRequestView view;
   private final HeaderPresenter headerPresenter;
   private ViewController viewParent;
+  private PopOverHelper popOverHelper;
 
   /**
    * Initializes a new RsccRequestPresenter with the matching view.
@@ -35,6 +36,7 @@ public class RsccRequestPresenter implements ControlledPresenter {
    */
   public void setViewParent(ViewController viewParent) {
     this.viewParent = viewParent;
+    popOverHelper = new PopOverHelper(viewParent, model);
   }
 
   private void attachEvents() {
@@ -79,5 +81,9 @@ public class RsccRequestPresenter implements ControlledPresenter {
       model.killConnection();
       viewParent.setView(RsccApp.HOME_VIEW);
     });
+    headerPresenter.setHelpBtnAction(event ->
+        popOverHelper.helpPopOver.show(view.headerView.helpBtn));
+    headerPresenter.setSettingsBtnAction(event ->
+    popOverHelper.settingsPopOver.show(view.headerView.settingsBtn));
   }
 }
