+++ src/ch/imedias/rsccfx/view/RsccSupportPresenter.java
@@ -18,6 +18,7 @@ public class RsccSupportPresenter implements ControlledPresenter {
   private final RsccSupportView view;
   private final HeaderPresenter headerPresenter;
   private ViewController viewParent;
+  private PopOverHelper popOverHelper;
 
   /**
    * Initializes a new RsccSupportPresenter with the according view.
@@ -42,6 +43,7 @@ public class RsccSupportPresenter implements ControlledPresenter {
    */
   public void setViewParent(ViewController viewParent) {
     this.viewParent = viewParent;
+    popOverHelper = new PopOverHelper(viewParent, model);
   }
 
   /**
@@ -107,6 +109,10 @@ public class RsccSupportPresenter implements ControlledPresenter {
   private void initHeader() {
     // Set all the actions regarding buttons in this method.
     headerPresenter.setBackBtnAction(event -> viewParent.setView("home"));
+    headerPresenter.setHelpBtnAction(event ->
+        popOverHelper.helpPopOver.show(view.headerView.helpBtn));
+    headerPresenter.setSettingsBtnAction(event ->
+        popOverHelper.settingsPopOver.show(view.headerView.settingsBtn));
     // TODO: Set actions on buttons (Help, Settings)
   }
